+++ Obsidian/Shared/WadExplorer.razor
@@ -1,0 +1,121 @@
+﻿@using LeagueToolkit.Core.Wad;
+@using Microsoft.WindowsAPICodePack.Dialogs;
+@using Obsidian.Data.Wad;
+@using Obsidian.Data;
+@using Obsidian.Services;
+@using Obsidian.Utils;
+@using PhotinoNET;
+@using Microsoft.AspNetCore.Components;
+
+<style>
+    .wad-tab-close-button {
+        border-radius: 0%;
+    }
+
+        .wad-tab-close-button:hover {
+            background-color: transparent !important;
+        }
+
+            .wad-tab-close-button:hover > .mud-icon-button-label > svg {
+                fill: #ff3f5fff;
+            }
+
+    .mud-tab {
+        text-transform: none;
+    }
+
+    .mud-tab-slider {
+        box-shadow: 0px 3px 5px -1px rgba(0,0,0,0.2),0px 5px 8px 0px rgba(0,0,0,0.14),0px 1px 14px 0px rgba(0,0,0,0.12);
+    }
+
+    .mud-tabs-panels {
+        background: rgb(58,59,71);
+        background: linear-gradient(0deg, rgba(58,59,71,1) 0%, rgba(50,51,61,1) 100%);
+    }
+</style>
+
+<MudPaper Style="height: 100%">
+    <MudPaper Elevation="25">
+        <CascadingValue Value="this">
+            <WadExplorerToolbar WadTree="WadTree"
+                                OnOpenWad="OpenWad"
+                                OnExtractAll="ExtractAll"
+                                OnExtractSelected="ExtractSelected"
+                                OnLoadHashtable="LoadHashtable" />
+        </CascadingValue>
+    </MudPaper>
+    <MudDivider />
+    <MudSplitter EnableSlide DisableMargin
+                 @ref="_splitter"
+                 Style="height: 100%;"
+                 Sensitivity="2.5f"
+                 Dimension="_splitterDimension" DimensionChanged="OnDimensionChanged">
+        <StartContent>
+            <MudStack Class="relative" Style="width: 100%"
+                      Spacing="1">
+                @if (WadTree is not null)
+                {
+                    <WadFilter FullWidth
+                           @ref="_wadFilterComponent"
+                           Filter="@WadTree.Filter" UseRegexFilter="@WadTree.UseRegexFilter"
+                           FilterChanged="OnFilterChanged" UseRegexFilterChanged="OnUseRegexFilterChanged" />
+                    <div class="mt-2" style="width: 100%">
+                        <TreeView TItem="WadTreeItemModel"
+                              ItemsFlat="@GetVisibleItemsForWadTree()"
+                              Height="calc(100vh - 119px)" ItemSize="36f">
+                            <ItemTemplate>
+                                <TreeWadItem Item="@context" Explorer="this" WadTree="WadTree" />
+                            </ItemTemplate>
+                        </TreeView>
+                    </div>
+                }
+                else
+                {
+                    <MudOverlay Absolute DarkBackground Visible ZIndex="9999">
+                        <MudProgressLinear Rounded Indeterminate
+                                       Class="my-6" Style="width: 100px"
+                                       Color="Color.Primary" Size="Size.Small" />
+                        <MudText Class="unselectable-text" Align="Align.Center" Typo="Typo.body2">
+                            Indexing...
+                        </MudText>
+                    </MudOverlay>
+                }
+            </MudStack>
+        </StartContent>
+        <EndContent>
+            @if (WadTree is not null)
+            {
+                @if (WadTree.CurrentPreviewType is WadFilePreviewType.Image)
+                {
+                    <WadFileImagePreview WadTree="WadTree" />
+                }
+                else if (WadTree.CurrentPreviewType is WadFilePreviewType.Viewport)
+                {
+                    <WadFileViewport WadTree="WadTree" />
+                }
+                <WadFileTextPreview @ref="WadTree.TextPreview" WadTree="WadTree" />
+            }
+        </EndContent>
+    </MudSplitter>
+
+    <MudOverlay DarkBackground Visible="_isLoadingWadFile" ZIndex="9999">
+        <MudProgressCircular Indeterminate Color="Color.Primary" />
+    </MudOverlay>
+    <MudOverlay DarkBackground Visible="_isExportingFiles" ZIndex="9999">
+        @*https://stackoverflow.com/a/61865580*@
+        <MudProgressLinear Rounded Indeterminate
+                           Class="my-6" Style="width: 500px"
+                           Color="Color.Primary" Size="Size.Large" />
+        <MudText Class="unselectable-text" Align="Align.Center" Typo="Typo.body1">
+            Extracting files...
+        </MudText>
+    </MudOverlay>
+    <MudOverlay DarkBackground Visible="_isLoadingHashtable" ZIndex="9999">
+        <MudProgressLinear Rounded Indeterminate
+                           Class="my-6" Style="width: 500px"
+                           Color="Color.Primary" Size="Size.Large" />
+    </MudOverlay>
+    <MudOverlay DarkBackground Visible="_isLoadingPreview" ZIndex="9999">
+        <MudProgressCircular Indeterminate Color="Color.Primary" />
+    </MudOverlay>
+</MudPaper>
